+++ utils/graphql-pydantic-converter/CHANGELOG.md
@@ -2,3 +2,8 @@
 - Add unit tests to project
 - Fix mutation input rendering for list items
 - Generate graphql schema request with custom typeOf depth
+
+# 0.1.0
+- Changed default payload boolean value from True to False
+- In query or mutation must be defined response properties
+- Generating of pydantic dataclasses for response parsing